+++ NEWS.md
@@ -1,3 +1,8 @@
+# gonovax 0.1.19
+
+* export data easily
+* add user function for running novax
+
 # gonovax 0.1.18
 
 * add mcmc thinning and sampling tools